Move pubkey recovery code to separate module
